+++ packages/fluid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @proteus-vue/fluid
 
+## 0.3.0-beta.11
+
+## 0.3.0-beta.10
+
+## 0.3.0-beta.9
+
+## 0.3.0-beta.8
+
 ## 0.3.0-beta.7
 
 ### Patch Changes