+++ packages/test-core/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @proteus-vue/test-core
 
+## 0.3.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @proteus-vue/compiler@0.3.0-beta.12
+
+## 0.3.0-beta.11
+
+### Patch Changes
+
+- @proteus-vue/compiler@0.3.0-beta.11
+
+## 0.3.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @proteus-vue/compiler@0.3.0-beta.10
+
+## 0.3.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @proteus-vue/compiler@0.3.0-beta.9
+
+## 0.3.0-beta.8
+
+### Patch Changes
+
+- @proteus-vue/compiler@0.3.0-beta.8
+
 ## 0.3.0-beta.7
 
 ### Patch Changes